Added Spring Cloud Bus
If there is any change in the GIT config value then this value will not reflected in the microservice until those instances are restarted.
To avoid this we can use the below endpoint to refresh the configuration from GIT.
/refresh
But this endpoint has to be performed on all the microservices instances which is difficult to maintain.So to avoid this spring cloud is provided.
when any microservice instance  receive a call from below endpoint
/bus/refresh
then spring cloud bus will trigger events to all the instances of that microservice to refresh so that config value is updated on all insances.
So if we provide /bus/refresh on one microservice instance spring cloud will automatically refresh other instances of that microservice instances.